pyproject.toml
[tool.poetry]
name = "aspire-app-v1-1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "aspire_app_v1"}]

[tool.poetry.dependencies]
python = "^3.11"
anyio = "3.7.1"
fastapi = {version = "0.115.3", extras = ["all"]}
fastapi-health = "^0.4.0"
python-dotenv = "^1.0.1"
sqlmodel = "^0.0.22"
toolz = "^1.0.0"
pysaml2 = "^7.5.0"
pytest = "^8.3.3"
starlette-context = "^0.3.6"
casbin = "^1.36.3"
canvasapi = "^3.3.0"
cryptography = "^43.0.1"
httpx = "^0.27.2"
langchain = "^0.3.3"
langchain-community = "^0.3.2"
langchain-core = "^0.3.10"
langchain-google-genai = "^2.0.1"
langchain-text-splitters = "^0.3.0"
langsmith = "^0.1.135"
loguru = "^0.7.2"
openai = "^1.51.2"
pandas = "^2.2.3"
psycopg2 = "^2.9.9"
pydantic = "^2.9.2"
pyjwt = "^2.9.0"
pymupdf = "^1.24.11"
pymupdfb = "^1.24.10"
requests = "^2.32.3"
rsa = "^4.9"
sqlalchemy = "^2.0.35"
uvicorn = "^0.31.1"
langchain-openai = "^0.2.5"
tiktoken = "^0.8.0"
pypdf = "^5.1.0"
python-magic = "^0.4.27"
langchain-huggingface = "^0.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
